refactor(spooky): extract app and listener group runtime
